+++ src/casino.h
@@ -2,13 +2,13 @@
 #define CASINO_H
 /* common declarations for the casino modules */
 
-extern integer bet;
-extern integer gld;
-extern integer tics;
+extern long bet;
+extern long gld;
+extern long tics;
 extern boolean c_closed;
 
 extern void c__display_gold();
-extern boolean c__get_response(vtype comment, integer *num);
+extern boolean c__get_response(vtype comment, long *num);
 extern void c__change_money();
 extern void c__check_casino_kickout();
